Nicolas Regnault: - add support for rational for fermions with LzMax>63 in FQHESphereJackGenerator
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@3207 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FQHE/src/HilbertSpace/FermionOnSphereLong.h
@@ -394,7 +394,7 @@ class FermionOnSphereLong :  public ParticleOnSphere
   // lzMax = momentum maximum value for a fermion
   // totalLz = momentum total value
   // return value = Hilbert space dimension
-  virtual int EvaluateHilbertSpaceDimension(int nbrFermions, int lzMax, int totalLz);
+  virtual long EvaluateHilbertSpaceDimension(int nbrFermions, int lzMax, int totalLz);
 
   // evaluate Hilbert space dimension with shifted values for lzMax and totalLz
   //
@@ -402,7 +402,7 @@ class FermionOnSphereLong :  public ParticleOnSphere
   // lzMax = two times momentum maximum value for a fermion plus one 
   // totalLz = momentum total value plus nbrFermions * (momentum maximum value for a fermion + 1)
   // return value = Hilbert space dimension  
-  int ShiftedEvaluateHilbertSpaceDimension(int nbrFermions, int lzMax, int totalLz);
+  long ShiftedEvaluateHilbertSpaceDimension(int nbrFermions, int lzMax, int totalLz);
 
   // generate look-up table associated to current Hilbert space
   //
